requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==7.1.2
colorama==0.4.4
ConfigArgParse==1.5.3
Flask==2.0.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.12.0
geventhttpclient==1.5.3
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
locust==2.7.3
MarkupSafe==2.0.1
msgpack==1.0.3
packaging==21.3
pluggy==1.0.0
psutil==5.9.0
py==1.11.0
pycparser==2.21
pyparsing==3.0.7
pytest==7.0.0
pytest-flask==1.2.0
python-dateutil==2.8.2
pywin32==303
pyzmq==22.3.0
requests==2.27.1
roundrobin==0.0.2
six==1.16.0
tomli==2.0.0
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.3
zope.event==4.5.0
zope.interface==5.4.0
